LF-5216 Use getDeviceType return to show 'Esc to close' hint conditionally; update Story; extract getDeviceType mock helpers
